Cuarta versión de la implementación de Clases
Estos archivos están asociados al issue [Estructura de clases realizada](https://github.com/NachoCarher/MyHams/issues/18) que avanza con el cumplimiento de [HU4](https://github.com/NachoCarher/MyHams/issues/13)
