+++ iv.yaml
@@ -1,4 +1,3 @@
 lenguaje: python
-entidades: 
-  -modules/productos.py
-  -modules/productosVenta.py
+entidad: modules/empresario.py
+
