+++ modules/productos.py
@@ -1,28 +1,21 @@
+# -*- coding: utf-8 -*-
+# 1)Clase para almacenar una lista de productos tipo "Producto" o "ProductoVenta"
+# 2)Gestiona los productos desde una abstracción superior por lo que la creación y eliminación de 
+# objetos tipo "Producto" o "ProductoVenta" debería hacerse desde aquí. 
 
-# Clase para almacenar todos los productos de un empresario (esten en venta o no)
 
 class Productos:
     def __init__(self):
-        self.productos= []	#El identificador de cada producto será su nombre
-        self.productosVenta = None	#Será un objeto de tipo "productosVenta"
-    
-    
+        self.productos= []	#Lista de objetos tipo "Producto" o "ProductoVenta"
+        
+     
     def getProductos(self):
         return self.productos
-    
-    def getProductosVenta(self):
-        return self.productosVenta
     
     def aniadirProducto(self,producto):
         pass
     
     def eliminarProducto(self,producto):
-        pass
-    
-    def aniadirProductoVenta(self,producto):
-        pass
-    
-    def eliminarProductoVenta(self,producto):
         pass
