--- modules/productosVenta.py
@@ -1,22 +1,0 @@
-
-#1) Lista de productos en Venta
-#2) Al igual que "producto", es manejado (creacion y eliminacion) desde "productos" y las operaciones especificas de productosVenta estarán aquí
-
-class productosVenta():
-    
-    def __init__(self,nom,info, cantidad, precio):
-        
-        self.productos = []
-
-    def getProductosVenta(self):
-        return self.productos
-        
-    def aniadirProductoVenta(self,producto):
-        pass
-    
-    def eliminarProductoVenta(self,producto):
-        pass
-    
-	
-
-
